decouple auto carousel from payment methods
